Port the code to use safe strings only
This patch introduces a strict separation between the mutable `bytes`
buffers and the immutable `string` objects, using the former to build
the latter when appropriate.

This should allow not to need the unsafe-strings flag anymore, which is
a good thing as this flag is on the way out in the Ocaml language.